+++ firebase-dataconnect/CHANGELOG.md
@@ -13,10 +13,10 @@
   must implement equals() to be a _logical_ comparsion, rather than just
   checking for _referencial_ equality using the `===` operator.
   ([#6424](https://github.com/firebase/firebase-android-sdk/pull/6424))
-* [feature] FirebaseDataConnectExperimental annotation added, and some
+* [feature] ExperimentalFirebaseDataConnect annotation added, and some
   APIs have been annotated with it, requiring applications that make use of
   these experimental APIs to opt-in using
-  `@OptIn(FirebaseDataConnectExperimental::class)` to suppress warnings or
+  `@OptIn(ExperimentalFirebaseDataConnect::class)` to suppress warnings or
   errors related to using these experimental APIs.
   ([#6424](https://github.com/firebase/firebase-android-sdk/pull/6424))